service

+++ src/main/java/com/banana/bananamint/persistence/ExpenseJPARepository.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface ExpenseJPARepository extends JpaRepository<Expense, Long> {
+public interface ExpenseJPARepository extends JpaRepository<Expense, Integer> {
     //public List<Expense> findAll(Long idCustomer) throws SQLException;
     //public List<Expense> findAllByDate(Long idCustomer, LocalDate initDate, LocalDate finalDate) throws SQLException;
     //public Expense save(Expense expense) throws SQLException;
